+++ packages/expo-updates/src/Updates.ts
@@ -22,7 +22,6 @@ export const updateId: string | null =
   ExpoUpdates.updateId && typeof ExpoUpdates.updateId === 'string'
     ? ExpoUpdates.updateId.toLowerCase()
     : null;
-export const releaseChannel: string = ExpoUpdates.releaseChannel ?? 'default';
 export const localAssets: LocalAssets = ExpoUpdates.localAssets ?? {};
 export const isEmergencyLaunch: boolean = ExpoUpdates.isEmergencyLaunch || false;
 export const isUsingEmbeddedAssets: boolean = ExpoUpdates.isUsingEmbeddedAssets || false;
@@ -32,6 +31,13 @@ if (ExpoUpdates.manifestString) {
   _manifest = JSON.parse(ExpoUpdates.manifestString);
 }
 export const manifest: Manifest | object = _manifest ?? {};
+
+// In Expo Go 2.19.x, the exported value ExpoUpdates.releaseChannel is always `default`, even if a
+// query parameter was used to load a published update from a non-default release channel. To fix
+// this for SDK 41 apps without patching native code, we can rely on the manifest.releaseChannel
+// property here if it exists, since all SDK 41 apps still use classic updates.
+export const releaseChannel: string =
+  _manifest?.releaseChannel ?? ExpoUpdates.releaseChannel ?? 'default';
 
 const isUsingDeveloperTool = !!(manifest as any).developer?.tool;
 const isUsingExpoDevelopmentClient = NativeModulesProxy.ExponentConstants?.appOwnership === 'expo';